Implemented working Sumo Wrestler code.
Sumo Wrestler sprite sheet is now in main.js and has working code in entity.js. Sumo does not bounce yet after landing. If you want to test it, you can edit scenemanager.js and uncomment:

{tag: "Sumo", x: 150, y: 200}, /* TODO: Sumo test */

which spawns it near Yamada in the beginning and test deflecting it with grappling hook and watch it bounce off walls in the beginning area.